Refactor: extracted stuff related to users from main controller
