+++ src/main/java/cmpt276/pg6/realtorest/controllers/UserController.java
@@ -1,0 +1,102 @@
+package cmpt276.pg6.realtorest.controllers;
+
+import java.util.List;
+import java.util.Map;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import cmpt276.pg6.realtorest.models.User;
+import cmpt276.pg6.realtorest.models.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+
+@Controller
+public class UserController {
+    @Autowired
+    private UserRepository userRepo;
+
+    public void setUserRepo(UserRepository userRepo) {
+        this.userRepo = userRepo;
+    }
+
+    /**
+     * Grabs the current URL and stores it as a model attribute, which means everything can use it. Mostly used for refreshing the page.
+     * Kevin: Note that this is something that is used in every controller, but I don't know how to extract this.
+     */
+    @ModelAttribute("currentUrl")
+    public String getCurrentUrl(HttpServletRequest request) {
+        return request.getRequestURI();
+    }
+
+    // Dev Page for Users Database
+    @GetMapping("/dev/users")
+    public String showDevPageUsers(Model model, HttpServletRequest request, HttpSession session) {
+        // Get all users from the database
+        List<User> users = userRepo.findAll();
+        model.addAttribute("users", users);
+        return "dev/users";
+    }
+
+    // Adding a user to the database, used for registering
+    @PostMapping("/users/add")
+    public String addUser(@RequestParam Map<String, String> newUser, HttpServletRequest request,
+        @RequestParam String redirectUrl, HttpServletResponse response, RedirectAttributes redirectAttributes) {
+        String username = newUser.get("username");
+        String email = newUser.get("email");
+        String password = newUser.get("password");
+        // Check if a user with the same email already exists
+
+        if (!userRepo.findByEmail(email).isEmpty()) {
+            redirectAttributes.addFlashAttribute("error",
+                "An account with this email already exists. Please try logging in.");
+            return "redirect:/login";
+        }
+        User user = new User(username, email, password);
+        userRepo.save(user);
+        request.getSession().setAttribute("session_user", user); // add user to session
+        response.setStatus(HttpServletResponse.SC_CREATED);
+        return "redirect:" + redirectUrl;
+    }
+
+    /**
+     * Fills the users database with testing data.
+     */
+    @PostMapping("/users/fill")
+    public String fillTestingDataUsers(@RequestParam String redirectUrl) {
+        userRepo.save(new User("Alice", "alice@email.com", "123"));
+        userRepo.save(new User("Bob", "bob@email.com", "456"));
+        userRepo.save(new User("Charlie", "charlie@email.com", "789"));
+        userRepo.save(new User("David", "david@email.com", "741"));
+        userRepo.save(new User("Eve", "eve@email.com", "852"));
+        userRepo.save(new User("Frank", "frank@email.com", "963"));
+        userRepo.save(new User("Grace", "grace@email.com", "846"));
+        userRepo.save(new User("Heidi", "heidi@email.com", "753"));
+        return "redirect:" + redirectUrl;
+    }
+
+    /**
+     * Deletes a user from the system.
+     */
+    @PostMapping("/users/delete/{uid}")
+    public String deleteUser(@PathVariable int uid, @RequestParam String redirectUrl) {
+        userRepo.deleteById(uid);
+        return "redirect:" + redirectUrl;
+    }
+
+    /**
+     * Deletes all users from the database.
+     * This is a dangerous operation and should not be used in a production environment.
+     */
+    @PostMapping("/users/delete/all")
+    public String deleteAllUsers(@RequestParam String redirectUrl) {
+        userRepo.deleteAll();
+        return "redirect:" + redirectUrl;
+    }
+}
